requirements3.txt
altgraph==0.10.2
bdist-mpkg==0.5.0
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
colorama==0.3.9
cycler==0.10.0
flake8==3.5.0
flake8-polyfill==1.0.2
idna==2.6
macholib==1.5.1
mando==0.6.4
matplotlib==2.0.2
mccabe==0.6.1
modulegraph==0.10.4
numpy==1.13.3
py2app==0.7.3
pycodestyle==2.3.1
pycparser==2.18
pycurl==7.43.0.1
pyflakes==1.6.0
pygit2==0.26.3
pyparsing==2.0.1
python-dateutil==2.6.1
pytz==2013.7
radon==2.1.1
requests==2.18.4
six==1.11.0
urllib3==1.22
